Re-factored the code by separating utility class
